<FeaturedTask/> component
